Agrupamento de métodos convertSnapshot() em um arquivo
