+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/repository/MedidaJuridicaRepository.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Map;
 
-import static com.uniprojecao.fabrica.gprojuridico.services.utils.Constants.MEDIDAS_JURIDICAS_COLLECTION;
-import static com.uniprojecao.fabrica.gprojuridico.services.utils.Utils.filterValidKeys;
+import static com.uniprojecao.fabrica.gprojuridico.utils.Constants.MEDIDAS_JURIDICAS_COLLECTION;
+import static com.uniprojecao.fabrica.gprojuridico.utils.Utils.filterValidKeys;
 
 @Repository
 @DependsOn("baseRepository")
