+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/services/exceptions/UserAlreadyCreatedException.java
@@ -1,6 +1,6 @@
 package com.uniprojecao.fabrica.gprojuridico.services.exceptions;
 
-import com.uniprojecao.fabrica.gprojuridico.services.utils.UsuarioUtils;
+import com.uniprojecao.fabrica.gprojuridico.services.DocumentSnapshotService;
 
 public class UserAlreadyCreatedException extends RuntimeException {
 
@@ -8,7 +8,7 @@ public class UserAlreadyCreatedException extends RuntimeException {
         super("Usuário com o email \"" + userEmail + "\" e CPF \"" + userCpf + "\" informados já existe.");
     }
 
-    public UserAlreadyCreatedException(UsuarioUtils.UserUniqueField field, String fieldValue) {
+    public UserAlreadyCreatedException(DocumentSnapshotService.UserUniqueField field, String fieldValue) {
         super("Usuário com o " + field.toString().toLowerCase() + " \"" + fieldValue + "\" já existe.");
     }
 }
